add clean code suggest to maintain SRP principle
