+++ Planilla_UCR/Application/Employees/IEmployeeService.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Domain.Employees.Entities;
+using Domain.People.Entities;
 
 namespace Application.Employees
 {
@@ -9,5 +10,7 @@ namespace Application.Employees
     {
         Task CreateEmployeeAsync(String email);
         Task<IEnumerable<Employee>> GetEmployeeByEmail(string email);
+        Task<IEnumerable<Person>> GetAllEmployees();
+        Task<IEnumerable<Person>> GetProjectEmployees(string projectName);
     }
 }
