+++ Planilla_UCR/Domain/Employees/Repositories/IEmployeeRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Domain.Employees.Entities;
+using Domain.People.Entities;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -9,5 +10,7 @@ namespace Domain.Employees.Repositories
     {
         Task CreateEmployeeAsync(String email);
         Task<IEnumerable<Employee>> GetEmployeeByEmail(string email);
+        Task<IEnumerable<Person>> GetAllEmployees();
+        Task<IEnumerable<Person>> GetProjectEmployees(string projectName);
     }
 }
