+++ Planilla_UCR/Domain/People/Repositories/IPersonRepository.cs
@@ -7,8 +7,6 @@ namespace Domain.People.Repositories
     public interface IPersonRepository
     {
         Task CreatePersonAsync(Person personInfo);
-        Task<IEnumerable<Person>> GetAllEmployees();
-        Task<IEnumerable<Person>> GetProjectEmployees(string projectName);
         Task<IEnumerable<Person>> GetPersonByEmail(string email);
         Task<IEnumerable<Person>> GetAllInfoEmployer(Person personInfo);
     }
