+++ Planilla_UCR/Infrastructure/Employees/EmployeeDbContext.cs
@@ -1,6 +1,8 @@
 ﻿using Domain.Employees.Entities;
+using Domain.People.Entities;
 using Infrastructure.Core;
 using Infrastructure.Employees.EntityMappings;
+using Infrastructure.People.EntityMappings;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 
@@ -13,11 +15,15 @@ namespace Infrastructure.Employees
         }
 
         public DbSet<Employee> Employees { get; set; } = null!;
+        public DbSet<Person> People { get; set; } = null!;
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             base.OnModelCreating(modelBuilder);
             modelBuilder.ApplyConfiguration(new EmployeeMap());
+
+            base.OnModelCreating(modelBuilder);
+            modelBuilder.ApplyConfiguration(new PersonMap());
         }
     }
 }
