+++ Planilla_UCR/Presentation/Employers/Components/HireEmployeesTable.razor
@@ -1,7 +1,8 @@
 ﻿@using Application.People
 @using Domain.People.Entities
+@using Application.Employees
 @inject NavigationManager nav
-@inject IPersonService PersonService
+@inject IEmployeeService PersonService
 
 @if(_loading) {
     <MudPaper  Height="220px" Width="100%" Outlined="false" Class="border-none pa-4" Elevation="0">
